+++ src/objects/label.js
@@ -1,6 +1,7 @@
 const utils = require("../utils/utils.js");
 const Objects = require('./objects.js');
-const THREE = require('./CSS2DRenderer.js');
+const CSS2D = require('./CSS2DRenderer.js');
+var THREE = require("../three.js");
 
 function Label(obj) {
 
@@ -8,12 +9,14 @@ function Label(obj) {
 
 	let div = Objects.prototype.drawLabelHTML(obj.htmlElement, obj.cssClass);
 
-	let label = new THREE.CSS2DObject(div);
+	let label = new CSS2D.CSS2DObject(div);
 	label.name = "label";
 	label.visible = obj.alwaysVisible;
 	label.alwaysVisible = obj.alwaysVisible;
-
-	var userScaleGroup = Objects.prototype._makeGroup(label, obj);
+	var projScaleGroup = new THREE.Group();
+	projScaleGroup.name = "scaleGroup";
+	projScaleGroup.add(label);
+	var userScaleGroup = Objects.prototype._makeGroup(projScaleGroup, obj);
 	Objects.prototype._addMethods(userScaleGroup);
 	userScaleGroup.visibility = obj.alwaysVisible;