+++ tests/threebox-tests-bundle.js
@@ -9405,7 +9405,10 @@ Threebox.prototype = {
 
 	// Objects
 
-	sphere: sphere,
+	sphere: function (options) {
+		options.bbox = this.options.enableSelectingObjects;
+		return sphere(options, this.world)
+	},
 
 	line: line,
 
@@ -9413,16 +9416,18 @@ Threebox.prototype = {
 
 	tooltip: tooltip,
 
-	tube: function (obj) {
-		return tube(obj, this.world)
+	tube: function (options) {
+		options.bbox = this.options.enableSelectingObjects;
+		return tube(options, this.world)
 	},
 
-	Object3D: function (obj, o) {
-		return Object3D(obj, o)
+	Object3D: function (options, o) {
+		options.bbox = this.options.enableSelectingObjects;
+		return Object3D(options, o)
 	},
 
 	loadObj: async function loadObj(options, cb) {
-
+		options.bbox = this.options.enableSelectingObjects;
 		//[jscastro] new added cache for 3D Objects
 		let cache = this.objectsCache.get(options.obj);
 		if (cache) {
@@ -9730,21 +9735,24 @@ Threebox.prototype = {
 	//[jscastro] method to fully dispose the resources, watch out is you call this without navigating to other page
 	dispose: async function () {
 
-		console.log(window.tb.memory());
+		console.log(this.memory());
 		//console.log(window.performance.memory);
 
-		return new Promise(disposed => {
-			this.clear(null, true);
-			this.map.remove();
-			this.map = {};
-			this.scene.remove(this.world);
-			this.scene.dispose();
-			this.world.children = [];
-			this.world = null;
-			this.labelRenderer.dispose();
-			console.log(window.tb.memory());
-			this.renderer.dispose();
-			disposed('dispose finished');
+		return new Promise((resolve) => {
+			resolve(
+				this.clear(null, true).then((resolve) => {
+					this.map.remove();
+					this.map = {};
+					this.scene.remove(this.world);
+					this.scene.dispose();
+					this.world.children = [];
+					this.world = null;
+					this.labelRenderer.dispose();
+					console.log(this.memory());
+					this.renderer.dispose();
+					return resolve;
+				})
+			);
 			//console.log(window.performance.memory);
 		});
 
@@ -10398,19 +10406,12 @@ THREE.CSS2DObject = function (element) {
 	this.dispose = function () {
 		this.remove();
 		this.element = null;
-		if (this.parent) this.parent.remove(this);
 	}
 
 	this.remove = function () {
-		this.traverse(function (object) {
-
-			if (object.element instanceof Element && object.element.parentNode !== null) {
-
-				object.element.parentNode.removeChild(object.element);
-
-			}
-
-		});
+		if (this.element instanceof Element && this.element.parentNode !== null) {
+			this.element.parentNode.removeChild(this.element);
+		}
 	}
 
 	this.addEventListener('removed', function () {
@@ -10889,7 +10890,8 @@ module.exports = exports = BuildingShadows;
 },{"../../utils/suncalc.js":96}],81:[function(require,module,exports){
 const utils = require("../utils/utils.js");
 const Objects = require('./objects.js');
-const THREE = require('./CSS2DRenderer.js');
+const CSS2D = require('./CSS2DRenderer.js');
+var THREE = require("../three.js");
 
 function Label(obj) {
 
@@ -10897,12 +10899,15 @@ function Label(obj) {
 
 	let div = Objects.prototype.drawLabelHTML(obj.htmlElement, obj.cssClass);
 
-	let label = new THREE.CSS2DObject(div);
+	let label = new CSS2D.CSS2DObject(div);
 	label.name = "label";
 	label.visible = obj.alwaysVisible;
 	label.alwaysVisible = obj.alwaysVisible;
-
-	var userScaleGroup = Objects.prototype._makeGroup(label, obj);
+	var projScaleGroup = new THREE.Group();
+	projScaleGroup.name = "group";
+	projScaleGroup.add(label);
+	var userScaleGroup = Objects.prototype._makeGroup(projScaleGroup, obj);
+	userScaleGroup.name = "object";
 	Objects.prototype._addMethods(userScaleGroup);
 	userScaleGroup.visibility = obj.alwaysVisible;
 
@@ -10911,7 +10916,7 @@ function Label(obj) {
 
 
 module.exports = exports = Label;
-},{"../utils/utils.js":97,"./CSS2DRenderer.js":76,"./objects.js":89}],82:[function(require,module,exports){
+},{"../three.js":93,"../utils/utils.js":97,"./CSS2DRenderer.js":76,"./objects.js":89}],82:[function(require,module,exports){
 var THREE = require("../three.js");
 var utils = require("../utils/utils.js");
 var Objects = require('./objects.js');
@@ -11992,32 +11997,20 @@ function loadObj(options, cb, promise) {
 			if (options.normalize) { normalizeSpecular(obj); }
 			obj.name = "model";
 			var projScaleGroup = new THREE.Group();
-			projScaleGroup.name = "group";
+			projScaleGroup.name = "scaleGroup";
 			projScaleGroup.add(obj)
 			var userScaleGroup = Objects.prototype._makeGroup(projScaleGroup, options);
-			userScaleGroup.name = "object";
-			//[jscastro] assign the animations to the userScaleGroup before enrolling it in AnimationsManager through _addMethods
-			userScaleGroup.animations = animations;
-
+			userScaleGroup.name = "threeboxObject";
 			Objects.prototype._addMethods(userScaleGroup);
 			//[jscastro] calculate automatically the pivotal center of the object
 			userScaleGroup.setAnchor(options.anchor);
 			//[jscastro] override the center calculated if the object has adjustments
 			userScaleGroup.setCenter(options.adjustment);
-
-			let anim = userScaleGroup.animations;
-
-			// [jscastro] after adding methods create the bounding box at userScaleGroup but add it to its children for positioning
-			let boxGrid = userScaleGroup.drawBoundingBox();
-			projScaleGroup.add(boxGrid);
-
-			//[jscastro] we add by default a tooltip that can be override later or hide it with threebox `enableTooltips`
-			userScaleGroup.addTooltip(userScaleGroup.uuid, true, userScaleGroup.anchor);
-
-			cb(userScaleGroup);
+			//[jscastro] return to cache
 			promise(userScaleGroup);
-
-			// [jscastro] initialize the default animation to avoid issues with position
+			//[jscastro] then return to the client-side callback
+			cb(userScaleGroup);
+			// [jscastro] initialize the default animation to avoid issues with skeleton position
 			userScaleGroup.idle();
 
 		}, () => (null), error => {
@@ -25137,6 +25130,18 @@ Objects.prototype = {
 
 			if (!obj.coordinates) obj.coordinates = [0, 0, 0];
 
+			//[jscastro] added property for the internal 3D model
+			Object.defineProperty(obj, 'model', {
+				get() { return obj.getObjectByName("model"); }
+			});
+
+			let _animations;
+			//[jscastro] added property for the internal 3D model
+			Object.defineProperty(obj, 'animations', {
+				get() { return obj.model.animations },
+				//set(value) { _animations = value}
+			});
+
 			// Bestow this mesh with animation superpowers and keeps track of its movements in the global animation queue			
 			root.animationManager.enroll(obj);
 
@@ -25166,7 +25171,7 @@ Objects.prototype = {
 				obj.set({ position: lnglat });
 				//Each time the object is positioned, set modelHeight property and project the floor
 				obj.modelHeight = obj.coordinates[2];
-				obj.setBoundingBoxShadowFloor();
+				if (obj.boxGroup) obj.setBoundingBoxShadowFloor();
 				return obj;
 
 			}
@@ -25238,19 +25243,33 @@ Objects.prototype = {
 				tb.map.repaint = true;
 			}
 
-			let _boundingBox;
+
+			//[jscastro] added property for scaled group inside threeboxObject
+			Object.defineProperty(obj, 'scaleGroup', {
+				get() {
+					return obj.getObjectByName("scaleGroup");
+				}
+			})
+
+			//[jscastro] added property for boundingBox group helper
+			Object.defineProperty(obj, 'boxGroup', {
+				get() {
+					return obj.getObjectByName("boxGroup");
+				}
+			})
+
 			//[jscastro] added property for boundingBox helper
 			Object.defineProperty(obj, 'boundingBox', {
 				get() {
-					return obj.getObjectByName("BoxModel");
+					return obj.getObjectByName("boxModel");
 				}
 			})
 
 			let _boundingBoxShadow;
-			//[jscastro] added property for boundingBox helper
+			//[jscastro] added property for boundingBox shadow helper
 			Object.defineProperty(obj, 'boundingBoxShadow', {
 				get() {
-					return obj.getObjectByName("BoxShadow");
+					return obj.getObjectByName("boxShadow");
 				}
 			})
 
@@ -25259,12 +25278,12 @@ Objects.prototype = {
 				//let's create 2 wireframes, one for the object and one to project on the floor position
 				let bb = obj.box3();
 				//create the group to return
-				let boxGrid = new THREE.Group();
-				boxGrid.name = "BoxGrid";
-				boxGrid.updateMatrixWorld(true);
+				let boxGroup = new THREE.Group();
+				boxGroup.name = "boxGroup";
+				boxGroup.updateMatrixWorld(true);
 				let boxModel = new THREE.Box3Helper(bb, Objects.prototype._defaults.colors.yellow);
-				boxModel.name = "BoxModel";
-				boxGrid.add(boxModel);
+				boxModel.name = "boxModel";
+				boxGroup.add(boxModel);
 				boxModel.layers.disable(0); // it makes the object invisible for the raycaster
 				//obj.boundingBox = boxModel;
 
@@ -25273,14 +25292,14 @@ Objects.prototype = {
 				//we make the second box flat and at the floor height level
 				bb2.max.z = bb2.min.z;
 				let boxShadow = new THREE.Box3Helper(bb2, Objects.prototype._defaults.colors.black);
-				boxShadow.name = "BoxShadow";
+				boxShadow.name = "boxShadow";
 
-				boxGrid.add(boxShadow);
+				boxGroup.add(boxShadow);
 				boxShadow.layers.disable(0); // it makes the object invisible for the raycaster
 				//obj.boundingBoxShadow = boxShadow;
 
-				boxGrid.visible = false; // visibility is managed from the parent
-				return boxGrid;
+				boxGroup.visible = false; // visibility is managed from the parent
+				return boxGroup;
 			}
 
 			//[jscastro] added method to position the shadow box on the floor depending the object height
@@ -25367,18 +25386,6 @@ Objects.prototype = {
 				get() { return obj.getObjectByName("tooltip"); }
 			});
 
-			//[jscastro] added property for the internal 3D model
-			Object.defineProperty(obj, 'model', {
-				get() { return obj.getObjectByName("model"); }
-			});
-
-			let _animations;
-			//[jscastro] added property for the internal 3D model
-			Object.defineProperty(obj, 'animations', {
-				get() { return _animations},
-				set(value) { _animations = value}
-			});
-
 			//[jscastro] added property to redefine visible, including the label and tooltip
 			Object.defineProperty(obj, 'visibility', {
 				get() { return obj.visible; },
@@ -25424,13 +25431,22 @@ Objects.prototype = {
 				}
 			}
 
+			//[jscastro] remove CSS2 label method 
+			obj.removeLabel = function () {
+				if (obj.label) {
+					obj.label.dispose();
+					let g = obj.children[0].children;
+					g.splice(g.indexOf(obj.label), 1);
+				}
+			}
+
 			//[jscastro] draw label method can be invoked separately
 			obj.drawLabelHTML = function (HTMLElement, visible = false, center = obj.anchor) {
 				let div = root.drawLabelHTML(HTMLElement, Objects.prototype._defaults.label.cssClass);
 				const box = obj.box3();
 				const size = box.getSize(new THREE.Vector3());
 				let bottomLeft = { x: box.max.x, y: box.max.y, z: box.min.z };
-				if (obj.label) { obj.label.remove; }
+				obj.removeLabel();
 				let label = new CSS2D.CSS2DObject(div);
 				label.name = "label";
 				label.position.set(((-size.x * 0.5) - obj.model.position.x - center.x + bottomLeft.x), ((-size.y * 0.5) - obj.model.position.y - center.y + bottomLeft.y), size.z * 0.5); //middle-centered
@@ -25447,13 +25463,22 @@ Objects.prototype = {
 					const box = obj.box3();
 					const size = box.getSize(new THREE.Vector3());
 					let bottomLeft = { x: box.max.x, y: box.max.y, z: box.min.z };
-					if (obj.tooltip) { obj.tooltip.remove; }
+					obj.removeTooltip();
 					let tooltip = new CSS2D.CSS2DObject(divToolTip);
 					tooltip.name = "tooltip";
 					tooltip.position.set(((-size.x * 0.5) - obj.model.position.x - center.x + bottomLeft.x), ((-size.y * 0.5) - obj.model.position.y - center.y + bottomLeft.y), size.z); //top-centered
 					tooltip.visible = false; //only visible on mouseover or selected
 					//we add it to the first children to get same boxing and position
 					obj.children[0].add(tooltip);
+				}
+			}
+
+			//[jscastro] remove CSS2 tooltip method
+			obj.removeTooltip = function () {
+				if (obj.tooltip) {
+					obj.tooltip.dispose();
+					let g = obj.children[0].children;
+					g.splice(g.indexOf(obj.tooltip), 1);
 				}
 			}
 
@@ -25579,6 +25604,10 @@ Objects.prototype = {
 				set(value) {
 					if (value) {
 						if (!obj.selected) {
+							if (obj.userData.bbox && !obj.boundingBox) {
+								obj.scaleGroup.add(obj.drawBoundingBox());
+							}
+
 							if (obj.boundingBox) {
 								obj.boundingBox.material = Objects.prototype._defaults.materials.boxOverMaterial;
 								obj.boundingBox.parent.visible = true;
@@ -25598,6 +25627,7 @@ Objects.prototype = {
 								obj.boundingBox.layers.disable(1);
 								obj.boundingBoxShadow.layers.disable(1);
 								obj.boundingBox.material = Objects.prototype._defaults.materials.boxNormalMaterial;
+								obj.remove(obj.boundingBox);
 							}
 							if (obj.label && !obj.label.alwaysVisible) { obj.label.visible = false; }
 						}
@@ -25668,27 +25698,66 @@ Objects.prototype = {
 		}
 
 		obj.add = function (o) {
-			obj.children[0].add(o);
+			obj.scaleGroup.add(o);
 			o.position.z = (obj.coordinates[2] ? -obj.coordinates[2] : 0);
 			return o;
 		}
 
 		obj.remove = function (o) {
-			obj.children[0].remove(o);
+			if (!o) return;
+			o.traverse(m => {
+				//console.log('dispose geometry!')
+				if (m.geometry) m.geometry.dispose();
+				if (m.material) {
+					if (m.material.isMaterial) {
+						cleanMaterial(m.material)
+					} else {
+						// an array of materials
+						for (const material of m.material) cleanMaterial(material)
+					}
+				}
+				if (m.dispose) m.dispose();
+			})
+
+			obj.scaleGroup.remove(o);
 			tb.map.repaint = true;
 		}
 
 		//[jscastro] clone + assigning all the attributes
 		obj.duplicate = function (options) {
-			let dupe = obj.clone(true);
-			dupe.userData = options || obj.userData;
-			root._addMethods(dupe);
-			dupe.deepCopy(obj);
 
-			return dupe;
+			let dupe = obj.clone(true);	//clone the whole threebox object
+			dupe.getObjectByName("model").animations = obj.animations; //we must set this explicitly before addMethods
+			if (dupe.userData.feature) dupe.userData.feature.properties.uuid = dupe.uuid;
+			root._addMethods(dupe); // add methods
+
+			if (!options || options.scale == obj.userData.scale) {
+				//no options, no changes, just return the same object
+				dupe.copyAnchor(obj); // copy anchors
+				//[jscastro] we add by default a tooltip that can be overriden later or hide it with threebox `enableTooltips`
+				return dupe;
+			} else {
+				dupe.userData = options;
+				dupe.userData.isGeoGroup = true;
+				dupe.remove(dupe.boxGroup);
+				// [jscastro] rotate and scale the model
+				const r = utils.types.rotation(options.rotation, [0, 0, 0]);
+				const s = utils.types.scale(options.scale, [1, 1, 1]);
+				// rotate and scale
+				dupe.model.rotation.set(r[0], r[1], r[2]);
+				dupe.model.scale.set(s[0], s[1], s[2]);
+				//[jscastro] calculate automatically the pivotal center of the object
+				dupe.setAnchor(options.anchor);
+				//[jscastro] override the center calculated if the object has adjustments
+				dupe.setCenter(options.adjustment);
+				return dupe;
+
+			}
+
 		}
 
-		obj.deepCopy = function (o) {
+		//[jscastro] copy anchor values
+		obj.copyAnchor = function (o) {
 
 			obj.anchor = o.anchor;
 			obj.none = { x: 0, y: 0, z: 0 };
@@ -25702,7 +25771,6 @@ Objects.prototype = {
 			obj.left = o.left;
 			obj.right = o.right;
 
-			return obj;
 		}
 
 		obj.dispose = function () {
@@ -25712,10 +25780,10 @@ Objects.prototype = {
 			obj.traverse(o => {
 				//don't dispose th object itself as it will be recursive
 				if (o.parent && o.parent.name == "world") return;
-				if (o.isMesh) {
-					//console.log('dispose geometry!')
-					o.geometry.dispose();
+				//console.log('dispose geometry!')
+				if (o.geometry) o.geometry.dispose();
 
+				if (o.material) {
 					if (o.material.isMaterial) {
 						cleanMaterial(o.material)
 					} else {
@@ -25854,7 +25922,8 @@ Objects.prototype = {
 			sides: 20,
 			units: 'scene',
 			material: 'MeshBasicMaterial',
-			anchor: 'bottom-left'
+			anchor: 'bottom-left',
+			bbox: false
 		},
 
 		label: {
@@ -25879,7 +25948,8 @@ Objects.prototype = {
 			sides: 6,
 			units: 'scene',
 			material: 'MeshBasicMaterial',
-			anchor: 'center'
+			anchor: 'center',
+			bbox: false
 		},
 
 		extrusion: {
@@ -25898,13 +25968,15 @@ Objects.prototype = {
 			scale: 1,
 			rotation: 0,
 			defaultAnimation: 0,
-			anchor: 'bottom-left'
+			anchor: 'bottom-left',
+			bbox: false
 		},
 
 		Object3D: {
 			obj: null,
 			units: 'scene',
-			anchor: 'bottom-left'
+			anchor: 'bottom-left',
+			bbox: false
 		}
 	},
 
@@ -25922,15 +25994,15 @@ var material = require("../utils/material.js");
 var Objects = require('./objects.js');
 var Object3D = require('./Object3D.js');
 
-function Sphere(options) {
+function Sphere(opt) {
 
-	options = utils._validate(options, Objects.prototype._defaults.sphere);
-	var geometry = new THREE.SphereBufferGeometry(options.radius, options.sides, options.sides);
-	var mat = material(options)
+	opt = utils._validate(opt, Objects.prototype._defaults.sphere);
+	var geometry = new THREE.SphereBufferGeometry(opt.radius, opt.sides, opt.sides);
+	var mat = material(opt)
 	var output = new THREE.Mesh(geometry, mat);
 
 	//[jscastro] we convert it in Object3D to add methods, bounding box, model, tooltip...
-	return new Object3D({ obj: output, units: options.units, anchor: options.anchor, adjustment: options.adjustment });
+	return new Object3D({ obj: output, units: opt.units, anchor: opt.anchor, adjustment: opt.adjustment, bbox: opt.bbox });
 
 }
 
@@ -25940,6 +26012,7 @@ module.exports = exports = Sphere;
 const utils = require("../utils/utils.js");
 const Objects = require('./objects.js');
 const CSS2D = require('./CSS2DRenderer.js');
+var THREE = require("../three.js");
 
 function Tooltip(obj) {
 
@@ -25952,7 +26025,11 @@ function Tooltip(obj) {
 		let tooltip = new CSS2D.CSS2DObject(divToolTip);
 		tooltip.visible = false;
 		tooltip.name = "tooltip";
-		var userScaleGroup = Objects.prototype._makeGroup(tooltip, obj);
+		var projScaleGroup = new THREE.Group();
+		projScaleGroup.name = "group";
+		projScaleGroup.add(tooltip);
+		var userScaleGroup = Objects.prototype._makeGroup(projScaleGroup, obj);
+		userScaleGroup.name = "object";
 		Objects.prototype._addMethods(userScaleGroup);
 		return userScaleGroup;
 	}
@@ -25960,31 +26037,31 @@ function Tooltip(obj) {
 }
 
 module.exports = exports = Tooltip;
-},{"../utils/utils.js":97,"./CSS2DRenderer.js":76,"./objects.js":89}],92:[function(require,module,exports){
+},{"../three.js":93,"../utils/utils.js":97,"./CSS2DRenderer.js":76,"./objects.js":89}],92:[function(require,module,exports){
 var utils = require("../utils/utils.js");
 var material = require("../utils/material.js");
 var Objects = require('./objects.js');
 var THREE = require("../three.js");
 var Object3D = require('./Object3D.js');
 
-function tube(obj, world){
+function tube(opt, world){
 
 	// validate and prep input geometry
-	var obj = utils._validate(obj, Objects.prototype._defaults.tube);
-    var straightProject = utils.lnglatsToWorld(obj.geometry);
+	var opt = utils._validate(opt, Objects.prototype._defaults.tube);
+    var straightProject = utils.lnglatsToWorld(opt.geometry);
 	var normalized = utils.normalizeVertices(straightProject);
 
-	var crossSection = tube.prototype.defineCrossSection(obj);
+	var crossSection = tube.prototype.defineCrossSection(opt);
 	var vertices = tube.prototype.buildVertices(crossSection, normalized.vertices, world);
-	var geom = tube.prototype.buildFaces(vertices, normalized.vertices, obj);
+	var geom = tube.prototype.buildFaces(vertices, normalized.vertices, opt);
 
-	var mat = material(obj);
+	var mat = material(opt);
 
     var mesh = new THREE.Mesh( geom, mat );
     //mesh.position.copy(normalized.position);
 
 	//[jscastro] we convert it in Object3D to add methods, bounding box, model, tooltip...
-	return new Object3D({ obj: mesh, units: obj.units, anchor: obj.anchor, adjustment: obj.adjustment });
+	return new Object3D({ obj: mesh, units: opt.units, anchor: opt.anchor, adjustment: opt.adjustment, bbox: opt.bbox });
 
 	//return mesh